Add provision module to keycloak-benchmark
Using docker and docker-compose add a starter docker-compose.yml to be able to bring up keycloak server and postgres db containers. The distribution used is a wildfly legacy server application and old store configurations are in place. In addition to that we have used cpusets within the docker-compose definitions to control the cpu quota allocated to the services within these containers. In future, we would like to add another setup for newer Quarkus based keycloak distribution and new store, both together or as separate definitions. This PR also includes the steps to deploy the  custom provider to the containerized Keycloak server and have the possibility to create seed data efficiently.

Closes #66