Adding Utility to create custom reports (fixes #100)
Utility downloads the needed binary. Generates the report taking the source simulation logs and a target report directory and writes the HTML|CSV report there. In the end, it removes the binaries. The utility can also optionally consume a custom mustache template used to customize the reports, which is currently hosted under benchmark/src/main/resources.

Closes #100